Matlab does not support char input to `vpa` anymore
`str2sym` needs to be applied to the first input of `vpa+  in Matlab.  On the other hand, Octave does not have `str2sym`, and allows char input to `vpa`. This is handled by the compiler by using `vpa(str2sym(...), ...)` in the compiled code but defining `str2sym` as the identity function in Octave