Correction to `Y$`
In August 2019 I introduced `str2sym` with `vpa` because MATLAB's `vpa` doesn't support char input anymore. But there was a bug: `str2sym` was applied to any input. Actually it should only applied to char input; not to numeric input. This is corrected now: `str2sym` is only applied to char input